+++ CHANGELOG.md
@@ -489,3 +489,4 @@
 - 2026-04-06: Docs campaign, 0001 initial, sender, chat, +3
 - 2026-04-06: Valid follow up, users, mutate queries, frontier, +3
 - 2026-04-08: Docs send, sender, embeddings, mod, +3
+- 2026-04-08: Features frontier, mutate queries, mod, campaigns, +3